Rename reader modules to avoid conflict with built in ones
